Remove tree-sitter-kotlin
Maintaining the tree-sitter-based grammar as primary syntax highlighter
was unfortunately more trouble than it's worth, due to the numerous
complications it introduces into the build system: The tree-sitter
grammar requires a native module compiled against the same node version
as Atom, frequently resulting in version mismatches and the inability to
compile on a newer node version. Additionally, supporting native modules
on Windows is a bit more complicated than on Unix-based systems, which
often ship a compiler toolchain as part of the distribution.

Using the TextMate grammar is, while not as accurate as the
tree-sitter-based grammar, vastly easier to support and more portable
across platforms.